requirements.txt
groq>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
typer>=0.9.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0
aiofiles>=23.0.0
